+++ cpp/docker/run_dictation_client_cpp.sh
@@ -8,7 +8,7 @@
 set -euo pipefail
 IFS=$'\n\t'
 
-IMAGE_VERSION=2.5.0
+IMAGE_VERSION=2.7.0
 
 SCRIPT=$(realpath "$0")
 SCRIPTPATH=$(dirname "${SCRIPT}")
@@ -33,10 +33,13 @@ Dictation ASR gRPC client options:
                         Wait for the service start for a given duration
                         in seconds. Additionally print service health status,
                         but only for a non-zero timeout value. (defaults to 0)
-  --streaming           If present, will perform asynchronous RPC. This is obligatory for audio content larger than 3.5 MB.
+  --sync                If present, will perform synchronous RPC instead of asynchronous (streaming) call. 
+                        It is not recommended to use this option for large files. For audio larger than 3.5MB, 
+                        recognition quality is degraded - for the best possible recognition, 
+                        send shorter audio fragments or use the streaming mode.
   --time-offsets        If true, returns also recognized word time offsets.
   --single-utterance    If set - the recognizer will detect a single spoken utterance.
-  --interim-results     If set - messages with temporal results will be shown.
+  --interim-results     If set - messages with interim results will be shown.
   --service-settings=SETTINGS
                         Semicolon-separated list of key=value pairs defining settings to be sent to service via gRPC request.
   --max-alternatives=MAX_ALTERNATIVES
@@ -66,8 +69,8 @@ while getopts "f:hs:-:" optchar; do
                 interim-results)
                     opts+=( "--interim-results=true" )
                     ;;
-                streaming)
-                    opts+=( "--streaming" )
+                sync)
+                    opts+=( "--sync" )
                     ;;
                 filename=*)
                     val=${OPTARG#*=}